+++ common/src/main/java/com/lowdragmc/lowdraglib/core/mixins/accessor/MouseHandlerAccessor.java
@@ -11,5 +11,6 @@ import org.spongepowered.asm.mixin.gen.Accessor;
  */
 @Mixin(MouseHandler.class)
 public interface MouseHandlerAccessor {
-    @Accessor int getActiveButton();
+    @Accessor( "activeButton")
+    int ldlib$getActiveButton();
 }