+++ common/src/main/java/com/lowdragmc/lowdraglib/gui/util/ClickData.java
@@ -26,7 +26,7 @@ public class ClickData {
     public ClickData() {
         MouseHandler mouseHelper = Minecraft.getInstance().mouseHandler;
         long id = Minecraft.getInstance().getWindow().getWindow();
-        this.button = mouseHelper instanceof MouseHandlerAccessor accessor ? accessor.getActiveButton() : -1;
+        this.button = mouseHelper instanceof MouseHandlerAccessor accessor ? accessor.ldlib$getActiveButton() : -1;
         this.isShiftClick = InputConstants.isKeyDown(id, GLFW.GLFW_KEY_LEFT_SHIFT) || InputConstants.isKeyDown(id, GLFW.GLFW_KEY_RIGHT_SHIFT);
         this.isCtrlClick = InputConstants.isKeyDown(id, GLFW.GLFW_KEY_LEFT_CONTROL) || InputConstants.isKeyDown(id, GLFW.GLFW_KEY_RIGHT_CONTROL);
         this.isRemote = true;