+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## v1.0.43
 * Fixed crash L2Hostility, which tried to access level during virtual chunk <init>.
+* Fixed crash with jGUI, they don't use mixin correctly